+++ HeThongQuanLyRapPhim/src/main/java/hcmute/edu/vn/HeThongQuanLyRapPhim/service/DiscountServiceImpl.java
@@ -1,18 +1,22 @@
 package hcmute.edu.vn.HeThongQuanLyRapPhim.service;
 
+import hcmute.edu.vn.HeThongQuanLyRapPhim.model.ChienDichGiamGia;
+import hcmute.edu.vn.HeThongQuanLyRapPhim.repository.DiscountCampaignRepository;
 import hcmute.edu.vn.HeThongQuanLyRapPhim.repository.DiscountRepository;
 import hcmute.edu.vn.HeThongQuanLyRapPhim.model.MaGiamGia;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 @Service
 public class DiscountServiceImpl implements DiscountService {
-    private DiscountRepository discountRepository;
+    private final DiscountRepository discountRepository;
+    private final DiscountCampaignRepository discountCampaignRepository;
+
     @Autowired
-    public DiscountServiceImpl(DiscountRepository discountRepository) {
+    public DiscountServiceImpl(DiscountRepository discountRepository, DiscountCampaignRepository discountCampaignRepository) {
         this.discountRepository = discountRepository;
+        this.discountCampaignRepository = discountCampaignRepository;
     }
 
     @Override
@@ -21,32 +25,37 @@ public class DiscountServiceImpl implements DiscountService {
     }
 
     @Override
+    public List<ChienDichGiamGia> findAllChienDichGiamGia() {
+        return discountCampaignRepository.findAll();
+    }
+
+    @Override
     public MaGiamGia findById(int theId) {
-//        Kiểu dữ liệu Optional giúp xử lý trường hợp dữ liệu có thể null một cách an toàn.
-        Optional<MaGiamGia> result = discountRepository.findById(theId);
-        MaGiamGia maGiamGia = null;
-        if (result.isPresent()) {
-            maGiamGia = result.get();
+        return discountRepository.findById(theId).orElse(null);
+    }
+
+    @Override
+    public MaGiamGia insert(MaGiamGia theMaGiamGia) {
+        return discountRepository.save(theMaGiamGia);
+    }
+
+    @Override
+    public MaGiamGia update(int id, MaGiamGia theMaGiamGia) {
+        MaGiamGia maGiamGia = discountRepository.findById(id).orElse(null);
+        if(maGiamGia != null) {
+            return discountRepository.save(theMaGiamGia);
         }
-        else {
-            throw new RuntimeException("Did not find combo bap nuoc id - " + theId);
+        return null;
+    }
+
+
+    @Override
+    public boolean deleteById(int theId) {
+        MaGiamGia maGiamGia = discountRepository.findById(theId).orElse(null);
+        if(maGiamGia != null) {
+            discountRepository.deleteById(theId);
+            return true;
         }
-
-        return maGiamGia;
-    }
-
-    @Override
-    public MaGiamGia save(MaGiamGia maGiamGia) {
-        return discountRepository.save(maGiamGia);
-    }
-
-    @Override
-    public void deleteById(int theId) {
-        discountRepository.deleteById(theId);
-    }
-
-    @Override
-    public MaGiamGia updateCustomer(MaGiamGia maGiamGia) {
-        return discountRepository.save(maGiamGia);
+        return false;
     }
 }
